renamed amber config packages
