+++ access-control-service/src/main/scala/org/apache/texera/service/AccessControlService.scala
@@ -23,7 +23,7 @@ import io.dropwizard.auth.AuthDynamicFeature
 import io.dropwizard.configuration.{EnvironmentVariableSubstitutor, SubstitutingSourceProvider}
 import io.dropwizard.core.Application
 import io.dropwizard.core.setup.{Bootstrap, Environment}
-import org.apache.texera.amber.config.StorageConfig
+import org.apache.texera.config.StorageConfig
 import org.apache.texera.auth.{
   JwtAuthFilter,
   RequestLoggingFilter,
