+++ amber/src/main/scala/org/apache/texera/amber/clustering/ClusterListener.scala
@@ -25,7 +25,7 @@ import org.apache.pekko.cluster.ClusterEvent._
 import com.google.protobuf.timestamp.Timestamp
 import com.twitter.util.{Await, Future}
 import org.apache.texera.amber.clustering.ClusterListener.numWorkerNodesInCluster
-import org.apache.texera.amber.config.ApplicationConfig
+import org.apache.texera.config.ApplicationConfig
 import org.apache.texera.amber.core.virtualidentity.ActorVirtualIdentity
 import org.apache.texera.amber.core.workflowruntimestate.FatalErrorType.EXECUTION_FAILURE
 import org.apache.texera.amber.core.workflowruntimestate.WorkflowFatalError
