+++ amber/src/main/scala/org/apache/texera/web/TexeraWebApplication.scala
@@ -26,7 +26,7 @@ import io.dropwizard.auth.AuthValueFactoryProvider
 import io.dropwizard.configuration.{EnvironmentVariableSubstitutor, SubstitutingSourceProvider}
 import io.dropwizard.setup.{Bootstrap, Environment}
 import io.dropwizard.websockets.WebsocketBundle
-import org.apache.texera.amber.config.StorageConfig
+import org.apache.texera.config.StorageConfig
 import org.apache.texera.amber.engine.common.Utils
 import org.apache.texera.amber.util.ObjectMapperUtils
 import org.apache.texera.auth.SessionUser
