+++ common/config/src/main/scala/org/apache/texera/config/StorageConfig.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.texera.amber.config
+package org.apache.texera.config
 
 import com.typesafe.config.{Config, ConfigFactory}
 import org.apache.texera.amber.util.ConfigParserUtil.parseSizeStringToBytes
