+++ common/workflow-core/src/main/scala/org/apache/texera/amber/util/IcebergUtil.scala
@@ -19,7 +19,7 @@
 
 package org.apache.texera.amber.util
 
-import org.apache.texera.amber.config.StorageConfig
+import org.apache.texera.config.StorageConfig
 import org.apache.texera.amber.core.tuple.{Attribute, AttributeType, LargeBinary, Schema, Tuple}
 import org.apache.hadoop.conf.Configuration
 import org.apache.iceberg.catalog.{Catalog, SupportsNamespaces, TableIdentifier}
