+++ common/workflow-core/src/test/scala/org/apache/texera/amber/storage/result/iceberg/IcebergDocumentSpec.scala
@@ -19,7 +19,7 @@
 
 package org.apache.texera.amber.storage.result.iceberg
 
-import org.apache.texera.amber.config.StorageConfig
+import org.apache.texera.config.StorageConfig
 import org.apache.texera.amber.core.state.State
 import org.apache.texera.amber.core.storage.model.{VirtualDocument, VirtualDocumentSpec}
 import org.apache.texera.amber.core.storage.{DocumentFactory, IcebergCatalogInstance, VFSURIFactory}
