+++ config-service/src/main/scala/org/apache/texera/service/ConfigService.scala
@@ -25,14 +25,13 @@ import io.dropwizard.auth.AuthDynamicFeature
 import io.dropwizard.configuration.{EnvironmentVariableSubstitutor, SubstitutingSourceProvider}
 import io.dropwizard.core.Application
 import io.dropwizard.core.setup.{Bootstrap, Environment}
-import org.apache.texera.amber.config.StorageConfig
 import org.apache.texera.auth.{
   JwtAuthFilter,
   RequestLoggingFilter,
   SessionUser,
   UnauthorizedExceptionMapper
 }
-import org.apache.texera.config.DefaultsConfig
+import org.apache.texera.config.{DefaultsConfig, StorageConfig}
 import org.apache.texera.dao.SqlServer
 import org.apache.texera.service.resource.{ConfigResource, HealthCheckResource}
 import org.eclipse.jetty.server.session.SessionHandler
