+++ file-service/src/test/scala/org/apache/texera/service/MockLakeFS.scala
@@ -21,7 +21,7 @@ package org.apache.texera.service
 
 import com.dimafeng.testcontainers._
 import io.lakefs.clients.sdk.{ApiClient, RepositoriesApi}
-import org.apache.texera.amber.config.StorageConfig
+import org.apache.texera.config.StorageConfig
 import org.apache.texera.service.util.S3StorageClient
 import org.scalatest.{BeforeAndAfterAll, Suite}
 import org.testcontainers.containers.Network
